Refactor InterceptorEndpointsRefresherScheduledService
